use size_t as type where useful (shut's up splint at many places)
svn path=/; revision=414

+++ include/base64.h
@@ -3,7 +3,7 @@
 
 #include "sstring.h"
 
-extern int b64decode(const unsigned char *, unsigned int, string *);
+extern int b64decode(const unsigned char *, size_t, string *);
 extern int b64encode(string *, string *);
 
 #endif
