+++ include/netio.h
@@ -2,18 +2,19 @@
 #define NETIO_H
 
 #include <bits/wordsize.h>
+#include <sys/types.h>
 
 extern char linein[];			/* current input line */
-extern unsigned int linelen;		/* length of the line */
+extern size_t linelen;			/* length of the line */
 
 extern int net_read(void);
 extern int net_writen(const char *const *);
 extern inline int netwrite(const char *);	/* same as net_write but expects that line is <= 512 characters
 						 * and includes <CRLF> */
-extern int netnwrite(const char *, const unsigned int);
+extern int netnwrite(const char *, const size_t);
 extern void ultostr(const unsigned long u, char *);
-extern int net_readbin(unsigned int, char *);
-extern int net_readline(unsigned int, char *);
+extern size_t net_readbin(size_t, char *);
+extern size_t net_readline(size_t, char *);
 extern int data_pending(void);
 
 static inline int
